+++ services/web/server/src/simcore_service_webserver/groups/_users.py
@@ -1,0 +1,15 @@
+"""
+NOTE: Coupling with user's plugin public modules should be added here to avoid cyclic dependencies
+"""
+
+from typing import Any, Mapping
+
+from ..users.schemas import convert_user_db_to_schema
+
+
+def convert_user_in_group_to_schema(user: Mapping[str, Any]) -> dict[str, str]:
+    group_user = convert_user_db_to_schema(user)
+    group_user.pop("role")
+    group_user["accessRights"] = user["access_rights"]
+    group_user["gid"] = user["primary_gid"]
+    return group_user
